requirementsL.txt
fastapi
gunicorn
mangum
line-bot-sdk
requests
